Class renaming and cleanup.
